refactor: ViewHolder 파일 분리
