+++ app/src/main/java/woowacourse/movie/presentation/ui/screen/adapter/ScreenRecyclerViewAdapter.kt
@@ -1,11 +1,7 @@
 package woowacourse.movie.presentation.ui.screen.adapter
 
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import android.widget.Button
-import android.widget.ImageView
-import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.RecyclerView.ViewHolder
 import woowacourse.movie.R
@@ -35,6 +31,7 @@ class ScreenRecyclerViewAdapter(
                 ScreenViewHolder(
                     LayoutInflater.from(parent.context)
                         .inflate(R.layout.holder_screen, parent, false),
+                    screenActionHandler,
                 )
 
             VIEW_TYPE_ADS ->
@@ -67,42 +64,6 @@ class ScreenRecyclerViewAdapter(
         screens.clear()
         screens.addAll(newScreens)
         notifyDataSetChanged()
-    }
-
-    inner class ScreenViewHolder(val view: View) : ViewHolder(view) {
-        private val poster: ImageView = view.findViewById(R.id.iv_poster)
-        private val title: TextView = view.findViewById(R.id.tv_title)
-        private val date: TextView = view.findViewById(R.id.tv_screen_date)
-        private val runningTime: TextView = view.findViewById(R.id.tv_screen_running_time)
-        private val reserveButton: Button = view.findViewById(R.id.btn_reserve_now)
-
-        fun bind(screen: Screen) {
-            initView(screen)
-            initClickListener(screen)
-        }
-
-        private fun initView(screen: Screen) {
-            with(screen) {
-                poster.setImageResource(movie.imageSrc)
-                title.text = movie.title
-                date.text = view.context.getString(R.string.screening_period, startDate, endDate)
-                runningTime.text = view.context.getString(R.string.running_time, movie.runningTime)
-            }
-        }
-
-        private fun initClickListener(screen: Screen) {
-            reserveButton.setOnClickListener {
-                screenActionHandler.onScreenClick(screen.id)
-            }
-        }
-    }
-
-    inner class AdsViewHolder(val view: View) : ViewHolder(view) {
-        private val poster: ImageView = view.findViewById(R.id.iv_ads)
-
-        fun bind(ads: Ads) {
-            poster.setImageResource(ads.poster)
-        }
     }
 
     companion object {
